Refactor and expose translation function.
No breaking changes on the external API. Capabilities -> Statement
function now exposed. All other changes are internal refactors.
