+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "capgrok"
 description = "Build wallet-signable messages with capability delegations that are human- and machine-readable."
 license = "Apache-2.0"
-version = "0.2.0"
+version = "0.2.1"
 edition = "2021"
 readme = "README.md"
 repository = "https://github.com/spruceid/capgrok/"
